Fix browser import session loss during database restore
- Add REST auth bypass for import endpoints using secret_key
- Use raw fetch() instead of wp.apiFetch() to avoid nonce middleware
- Preserve and restore user session_tokens during database import
- Import files before database, critical tables (options/usermeta) last
- Add chunked URL replacement with progress tracking

+++ src/Archive/SwishExtractor.php
@@ -293,7 +293,7 @@ class SwishExtractor extends SwishArchiver {
 				$dest_path,
 				$file_offset,
 				$bytes,
-				$timeout_seconds - ( microtime( true ) - $start_time )
+				(int) ( $timeout_seconds - ( microtime( true ) - $start_time ) )
 			);
 
 			if ( ! $completed ) {
